refactor(cli): extract loop session layer + extended state layer (CORE-2, option b+c)
Option (b): move isStalled/isRetryStorm/checkStopConditions/printSummary
from scripts/loop.js into scripts/lib/loop-state.js; loop.js re-exports
isStalled/isRetryStorm/checkStopConditions so the existing test surface
is unchanged.

Option (c): new scripts/lib/loop-session.js owning the claude -p session
layer (extractCost, spawnSession, spawnSessionAsync); loop.js imports and
re-exports spawnSession/spawnSessionAsync. buildTaskPrompt and
parseLoopArgs/printLoopHelp stay in loop.js (owned by AF-4 and AF-8/AF-10).

Acceptance evidence:
1. wc -l scripts/loop.js = 467 (gate: <= 467)
2. npx jest tests/scripts/loop.test.js: 30 passed, 0 failed;
   git diff main...HEAD -- tests/scripts/loop.test.js is empty
3. Legacy .arcforge-loop.json loads: loop-state.test.js legacy case +
   node -e loadLoopState() on a hand-written legacy file (iteration 7,
   pattern dag, cost 1.25 round-trip intact)
4. npm run lint:fix (no fixes applied) + npm run lint green
5. npm test all 5 runners green: Jest 2018/2018, hooks 268 pass / 0 fail,
   node runner all pass, pytest 541 passed, observer-daemon 16/16

Pre-existing TS-LSP diagnostics left untouched (also on main):
break after process.exit(0) in parseLoopArgs; unused _probe in
tryCreateCoordinator.
